Relocate API Types to Separate Module
