+++ pom.xml
@@ -55,6 +55,7 @@
 
 	<modules>
 		<module>vcell-admin</module>
+		<module>vcell-api-types</module>
 		<module>vcell-api</module>
 		<module>vcell-apiclient</module>
 		<module>vcell-client</module>
