+++ vcell-api-types/src/main/java/org/vcell/api/types/common/SimulationTaskRepresentation.java
@@ -1,4 +1,4 @@
-package cbit.rmi.event.client.common;
+package org.vcell.api.types.common;
 
 
 public class SimulationTaskRepresentation {
