+++ vcell-api-types/src/main/java/org/vcell/api/types/events/ExportVariableSpecs.java
@@ -1,4 +1,4 @@
-package cbit.rmi.event.client.events;
+package org.vcell.api.types.events;
 
 public class ExportVariableSpecs {
 	public final String[] variableNames;
