+++ vcell-api/src/main/java/org/vcell/rest/events/RestEventService.java
@@ -1,32 +1,19 @@
 package org.vcell.rest.events;
 
+import cbit.rmi.event.*;
+import cbit.vcell.message.VCMessagingService;
+import com.google.gson.Gson;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.vcell.api.types.events.*;
+import org.vcell.rest.server.ClientTopicMessageCollector;
+import org.vcell.util.Compare;
+import org.vcell.api.types.utils.DTOOldAPI;
+
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.concurrent.ConcurrentLinkedDeque;
 import java.util.concurrent.atomic.AtomicLong;
-
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-import cbit.rmi.event.client.events.BroadcastEventRepresentation;
-import cbit.rmi.event.client.events.DataJobEventRepresentation;
-import cbit.rmi.event.client.events.EventWrapper;
-import cbit.rmi.event.client.events.EventWrapper.EventType;
-import cbit.rmi.event.client.events.ExportEventRepresentation;
-import cbit.rmi.event.client.events.SimulationJobStatusEventRepresentation;
-import cbit.rmi.event.client.DTOOldAPI;
-import org.vcell.rest.server.ClientTopicMessageCollector;
-import org.vcell.util.Compare;
-
-import com.google.gson.Gson;
-
-import cbit.rmi.event.DataJobEvent;
-import cbit.rmi.event.ExportEvent;
-import cbit.rmi.event.MessageEvent;
-import cbit.rmi.event.PerformanceMonitorEvent;
-import cbit.rmi.event.SimulationJobStatusEvent;
-import cbit.rmi.event.VCellMessageEvent;
-import cbit.rmi.event.WorkerEvent;
-import cbit.vcell.message.VCMessagingService;
 
 public class RestEventService {
 	private final static Logger lg = LogManager.getLogger(RestEventService.class);
@@ -45,7 +32,7 @@ public class RestEventService {
 		}
 	}
 		
-	public void insert(String userid, EventType eventType, String eventJSON) {
+	public void insert(String userid, EventWrapper.EventType eventType, String eventJSON) {
 		long id = eventSequence.getAndIncrement();
 		long timestamp = System.currentTimeMillis();
 		EventWrapper wrapper = new EventWrapper(id, timestamp, userid, eventType, eventJSON);
@@ -67,7 +54,7 @@ public class RestEventService {
 				}
 				Gson gson = new Gson();
 				String eventJSON = gson.toJson(exportEventRep);
-				insert(exportEventRep.username,EventType.ExportEvent,eventJSON);
+				insert(exportEventRep.username, EventWrapper.EventType.ExportEvent,eventJSON);
 			}catch (Exception e) {
 				lg.error(e.getMessage(), e);
 			}
@@ -84,7 +71,7 @@ public class RestEventService {
 				}
 				Gson gson = new Gson();
 				String eventJSON = gson.toJson(simJobEventRep);
-				insert(simJobEventRep.username,EventType.SimJob,eventJSON);
+				insert(simJobEventRep.username, EventWrapper.EventType.SimJob,eventJSON);
 			}catch (Exception e) {
 				lg.error(e.getMessage(), e);
 			}
@@ -107,7 +94,7 @@ public class RestEventService {
 				//Add new broadcast message
 				Gson gson = new Gson();
 				String eventJSON = gson.toJson(broadcastEventRepresentation);
-				insert(null,EventType.Broadcast,eventJSON);
+				insert(null, EventWrapper.EventType.Broadcast,eventJSON);
 			}else {
 				lg.error("event of type VCellMessageEvent:"+vcellMessageEvent.getEventTypeID()+" not supported");
 			}
@@ -130,7 +117,7 @@ public class RestEventService {
 				}
 				Gson gson = new Gson();
 				String eventJSON = gson.toJson(dataJobEventRep);
-				insert(dataJobEventRep.username,EventType.DataJob,eventJSON);
+				insert(dataJobEventRep.username, EventWrapper.EventType.DataJob,eventJSON);
 			}catch (Exception e) {
 				lg.error(e.getMessage(), e);
 			}
