+++ vcell-api/src/main/java/org/vcell/rest/health/HealthService.java
@@ -14,12 +14,11 @@ import cbit.vcell.solver.VCSimulationIdentifier;
 import cbit.vcell.xml.XMLSource;
 import cbit.vcell.xml.XmlHelper;
 import com.google.gson.Gson;
+import org.vcell.api.types.events.EventWrapper;
+import org.vcell.api.types.events.SimulationJobStatusEventRepresentation;
 import org.apache.commons.io.IOUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import cbit.rmi.event.client.events.EventWrapper;
-import cbit.rmi.event.client.events.SimulationJobStatusEventRepresentation;
-import cbit.rmi.event.client.DTOOldAPI;
 import org.vcell.rest.events.RestEventService;
 import org.vcell.rest.server.RestDatabaseService;
 import org.vcell.util.BigString;
@@ -27,6 +26,7 @@ import org.vcell.util.document.KeyValue;
 import org.vcell.util.document.User;
 import org.vcell.util.document.UserInfo;
 import org.vcell.util.document.UserLoginInfo;
+import org.vcell.api.types.utils.DTOOldAPI;
 
 import java.util.ArrayList;
 import java.util.concurrent.ConcurrentLinkedDeque;
