+++ vcell-api/src/main/java/org/vcell/rest/users/NewUserRestlet.java
@@ -11,13 +11,13 @@ import org.restlet.Restlet;
 import org.restlet.data.*;
 import org.restlet.ext.json.JsonRepresentation;
 import org.restlet.representation.Representation;
-import cbit.rmi.event.client.DTOOldAPI;
 import org.vcell.rest.VCellApiApplication;
 import org.vcell.util.BeanUtils;
 import org.vcell.util.DataAccessException;
 import org.vcell.util.UseridIDExistsException;
 import org.vcell.util.document.UserInfo;
 import org.vcell.util.document.UserLoginInfo.DigestedPassword;
+import org.vcell.api.types.utils.DTOOldAPI;
 
 import java.io.IOException;
 import java.sql.SQLException;
@@ -33,7 +33,7 @@ public final class NewUserRestlet extends Restlet {
 	private void handleJsonRequest(Request request, Response response) {
 		String content = request.getEntityAsText();
 		Gson gson = new Gson();
-		cbit.rmi.event.client.common.UserInfo userinfo = gson.fromJson(content, cbit.rmi.event.client.common.UserInfo.class);
+		org.vcell.api.types.common.UserInfo userinfo = gson.fromJson(content, org.vcell.api.types.common.UserInfo.class);
 		
 		if (userinfo.email.length()<4){
 			response.setStatus(Status.CLIENT_ERROR_FORBIDDEN);
@@ -64,7 +64,7 @@ public final class NewUserRestlet extends Restlet {
 			VCellApiApplication vcellApiApplication = (VCellApiApplication)getApplication();
 			try {
 				UserInfo insertedUserInfo = vcellApiApplication.getRestDatabaseService().addUser(newUserInfo);
-				cbit.rmi.event.client.common.UserInfo inserted = DTOOldAPI.getApiUserInfo(insertedUserInfo);
+				org.vcell.api.types.common.UserInfo inserted = DTOOldAPI.getApiUserInfo(insertedUserInfo);
 				String userInfoJson = gson.toJson(inserted);
 				JsonRepresentation userRep = new JsonRepresentation(userInfoJson);
 				response.setStatus(Status.SUCCESS_CREATED);
