+++ vcell-apiclient/src/main/java/org/vcell/api/client/VCellApiClient.java
@@ -2,9 +2,13 @@ package org.vcell.api.client;
 
 import cbit.rmi.event.VCellApiRpcBody;
 import cbit.rmi.event.VCellApiRpcRequest;
-import cbit.rmi.event.client.DTOOldAPI;
 import com.google.gson.Gson;
 import com.nimbusds.oauth2.sdk.ParseException;
+import org.vcell.api.types.common.BiomodelRepresentation;
+import org.vcell.api.types.common.SimulationRepresentation;
+import org.vcell.api.types.common.SimulationTaskRepresentation;
+import org.vcell.api.types.common.UserInfo;
+import org.vcell.api.types.events.EventWrapper;
 import org.apache.http.*;
 import org.apache.http.client.ClientProtocolException;
 import org.apache.http.client.HttpResponseException;
@@ -28,11 +32,6 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.vcell.api.client.query.BioModelsQuerySpec;
 import org.vcell.api.client.query.SimTasksQuerySpec;
-import cbit.rmi.event.client.common.BiomodelRepresentation;
-import cbit.rmi.event.client.common.SimulationRepresentation;
-import cbit.rmi.event.client.common.SimulationTaskRepresentation;
-import cbit.rmi.event.client.common.UserInfo;
-import cbit.rmi.event.client.events.EventWrapper;
 import org.vcell.restclient.ApiClient;
 import org.vcell.restclient.ApiException;
 import org.vcell.restclient.CustomApiClientCode;
@@ -41,6 +40,7 @@ import org.vcell.restclient.api.UsersResourceApi;
 import org.vcell.restclient.auth.InteractiveLogin;
 import org.vcell.restclient.model.AccesTokenRepresentationRecord;
 import org.vcell.restclient.model.UserIdentityJSONSafe;
+import org.vcell.api.types.utils.DTOOldAPI;
 
 import java.awt.*;
 import java.io.*;
