+++ vcell-apiclient/src/main/java/org/vcell/api/client/examples/VCellApiClientTest.java
@@ -2,13 +2,14 @@ package org.vcell.api.client.examples;
 
 import java.io.IOException;
 
+import org.vcell.api.types.common.ApplicationRepresentation;
+import org.vcell.api.types.common.BiomodelRepresentation;
+import org.vcell.api.types.common.SimulationRepresentation;
+import org.vcell.api.types.common.SimulationTaskRepresentation;
 import org.vcell.api.client.VCellApiClient;
 import org.vcell.api.client.query.BioModelsQuerySpec;
 import org.vcell.api.client.query.SimTasksQuerySpec;
-import cbit.rmi.event.client.common.ApplicationRepresentation;
-import cbit.rmi.event.client.common.BiomodelRepresentation;
-import cbit.rmi.event.client.common.SimulationRepresentation;
-import cbit.rmi.event.client.common.SimulationTaskRepresentation;
+
 
 public class VCellApiClientTest {
