+++ vcell-apiclient/src/test/java/org/vcell/api/messaging/RemoteProxyVCellConnectionFactoryTest.java
@@ -2,10 +2,10 @@ package org.vcell.api.messaging;
 
 import cbit.vcell.message.server.bootstrap.client.RemoteProxyException;
 import cbit.vcell.server.VCellConnection;
+import org.vcell.api.types.common.BiomodelRepresentation;
 import org.junit.jupiter.api.*;
 import org.vcell.api.client.VCellApiClient;
 import org.vcell.api.client.query.BioModelsQuerySpec;
-import cbit.rmi.event.client.common.BiomodelRepresentation;
 import org.vcell.util.DataAccessException;
 import org.vcell.util.document.UserLoginInfo;
 import org.vcell.util.document.VCInfoContainer;
