+++ vcell-client/pom.xml
@@ -155,7 +155,13 @@
 			<artifactId>picocli</artifactId>
 			<version>${picocli.version}</version>
 		</dependency>
-	</dependencies>
+        <dependency>
+            <groupId>org.vcell</groupId>
+            <artifactId>vcell-api-types</artifactId>
+            <version>1.0.0</version>
+            <scope>compile</scope>
+        </dependency>
+    </dependencies>
 	<profiles>
 		<profile>
 			<id>build-documentation</id>
